requirements.txt
alembic==1.11.0
asgiref==3.6.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==3.0.10
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
flask-openapi3==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
Werkzeug==2.3.4
